do process newly created records
when the updated_time==0 the record has not been updated since creation.
Imagine the following scenario:

1. someone updates a record (updated_time != 0).
2. someone creates a new record (updated_time == 0, but created_time is
   larger than (1)).
3. run the exporter  and crash after the newly created (1) item is seen
4. execute again and (1) is newer processed.

Instead, we process if both the created_time and updated_time is
larger than the last checkpoint. There is just no way to process items
in order given the way the API is done